requirements.txt
djangorestframework==3.13.1
Django~=4.1
Pygments~=2.12.0
http==0.02